refactor(backend-user): update service, controller and user response dto

+++ backend/src/controllers/user.controller.ts
@@ -1,7 +1,9 @@
 import { Response } from "express";
+
 import { UserService } from "../services/user.service";
 import { UserUpdateDTO } from "../dtos/user/user-update.dto";
 import { AuthRequest } from "../middlewares/auth.middleware.js";
+import { UserResponseDTO } from "../dtos/user/user-response.dto";
 
 export class UserController {
   private userService: UserService;
@@ -10,38 +12,54 @@ export class UserController {
     this.userService = new UserService();
   }
 
-  async getMe(req: AuthRequest, res: Response) {
+  async getMe(req: AuthRequest, res: Response): Promise<Response> {
     try {
-      const userId = (req.user as any).id;
-      const user = await this.userService.getByIdWithPassword(userId);
-      if (!user) return res.status(404).json({ message: "User not found" });
+      const userId: string | undefined = req.user?.id;
+      if (!userId) 
+        return res.status(401).json({ message: "Unauthorized." });
 
-      res.json(user);
+      const userDTO: UserResponseDTO | null = await this.userService.getByIdWithPassword(userId);
+      if (!userDTO) 
+        return res.status(404).json({ message: "User not found." });
+
+      return res.status(200).json(userDTO);
     } catch (error: any) {
-      res.status(500).json({ error: error.message });
+      return res.status(500).json({ message: "Internal server error.", error: error.message });
     }
   }
 
-  async update(req: AuthRequest, res: Response) {
+  async update(req: AuthRequest, res: Response): Promise<Response> {
     try {
-      const userId = (req.user as any).id;
+      const userId: string | undefined = req.user?.id;
+      if (!userId) 
+        return res.status(401).json({ message: "Unauthorized." });
+
       const updateData: UserUpdateDTO = req.body;
-      const updatedUser = await this.userService.update(userId, updateData);
-      if (!updatedUser) return res.status(404).json({ message: "User not found" });
-      res.json(updatedUser);
+      const userDTO: UserResponseDTO | null = await this.userService.update(userId, updateData);
+
+      if (!userDTO) 
+        return res.status(404).json({ message: "User not found" });
+
+      return res.status(200).json(userDTO);
     } catch (error: any) {
-      res.status(500).json({ error: error.message });
+      return res.status(500).json({ message: "Internal server error.", error: error.message });
     }
   }
 
-  async delete(req: AuthRequest, res: Response) {
-    try {
-      const userId = (req.user as any).id;
-      const deletedUser = await this.userService.delete(userId);
-      if (!deletedUser) return res.status(404).json({ message: "User not found" });
-      res.json({ message: "User deactivated successfully" });
-    } catch (error: any) {
-      res.status(500).json({ error: error.message });
-    }
+  async delete(req: AuthRequest, res: Response): Promise<Response> {
+  try {
+    const userId: string | undefined = req.user?.id;
+    if (!userId) 
+      return res.status(401).json({ message: "Unauthorized." });
+
+    const deletedUser = await this.userService.delete(userId);
+    if (!deletedUser) 
+      return res.status(404).json({ message: "User not found." });
+
+    return res.status(204).send();
+  } catch (error: any) {
+    return res.status(500).json({ message: "Internal server error.", error: error.message });
   }
+}
+
 }
